Nginx Fixes + New easier way to run the toolkit + new build badge!

+++ inc/app-folders-create.sh
@@ -2,7 +2,13 @@
 echo
 sleep 1
 echo -e "$YELLOW--->Creating $APPTITLE files / folders...$ENDCOLOR"
-cp $SCRIPTPATH/$APPNAME/$APPNAME-folders-list $SCRIPTPATH/tmp/$APPSHORTNAME-folders-list || { echo -e "${RED}Creating folders list failed.$ENDCOLOR"; exit 1; }
+
+if [[ -f $SCRIPTPATH/utils/$APPNAME/$APPNAME-folders-list ]]; then
+   cp $SCRIPTPATH/utils/$APPNAME/$APPNAME-folders-list $SCRIPTPATH/tmp/$APPSHORTNAME-folders-list || { echo -e "${RED}Creating folders list failed.$ENDCOLOR"; exit 1; }
+else
+   cp $SCRIPTPATH/$APPNAME/$APPNAME-folders-list $SCRIPTPATH/tmp/$APPSHORTNAME-folders-list || { echo -e "${RED}Creating folders list failed.$ENDCOLOR"; exit 1; }
+fi
+
 sudo sed -i "s@APPPATH@$APPPATH@g" $SCRIPTPATH/tmp/$APPSHORTNAME-folders-list  || { echo -e "${RED}Replacing $APPTITLE path in folders list failed.$ENDCOLOR"; exit 1; }
 sudo sed -i "s@APPSETTINGSDIR@$APPSETTINGSDIR@g" $SCRIPTPATH/tmp/$APPSHORTNAME-folders-list  || { echo -e "${RED}Replacing $APPTITLE path in folders list failed.$ENDCOLOR"; exit 1; }
 sudo sed -i "s@UNAME@$UNAME@g" $SCRIPTPATH/tmp/$APPSHORTNAME-folders-list  || { echo -e "${RED}Replacing username in folders list failed.$ENDCOLOR"; exit 1; }
